Changed check in AddBook method to use long.Tryparse instead of int to account for ISBN numbers being bigger numbers than int is built for

+++ Labb4-EnhetsTestning/LibrarySystem.cs
@@ -30,7 +30,7 @@ namespace Labb4_EnhetsTestning
             {
                 return false;
             }
-            else if(!int.TryParse(book.ISBN, out int result))
+            else if(!long.TryParse(book.ISBN, out long result))
             {
                 return false;
             }
